CI fix

+++ include/box2d/types.h
@@ -17,7 +17,7 @@
 #define B2_LITERAL(T) (T)
 #endif
 
-#define B2_ARRAY_COUNT(A) (sizeof(A) / sizeof(A[0]))
+#define B2_ARRAY_COUNT(A) (int)(sizeof(A) / sizeof(A[0]))
 #define B2_MAYBE_UNUSED(x) ((void)(x))
 #define B2_NULL_INDEX (-1)
